moved reusable card component to dashboard dir
