+++ app/javascript/src/components/Dashboard/Notes/ListNoteCards.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import Card from "components/Card";
+import Card from "components/Dashboard/Card";
 
 import { calculateAge, formatTime } from "./utils/prettyPrintTime";
